Get Total Liked Recipe Count

+++ SpoonacularConcept/Models/ViewModels/AddToLikeCart.cs
@@ -10,8 +10,8 @@ namespace SpoonacularConcept.Models.ViewModels
         public string Image { get; set; }
         public string Summary { get; set; }
         public string Title { get; set; }
-        public int Servings { get; set; }
-        public int Score { get; set; }
+        public Int16 Servings { get; set; }
+        public byte Score { get; set; }
         public float Price { get; set; }
         public string Time { get; set; }
         public int recipeId { get; set; }
